+++ src/pages/CompanyPortal.tsx
@@ -2,9 +2,7 @@ import {
     Header,
     Main, 
     Container, 
-    Footer, 
     Row,
-    Grid,
     Copyright,
     ProfilePicture,
     UserButton,
@@ -13,22 +11,13 @@ import {
     ArrowPicture,
     HeaderLogo,
     Position, 
-    FacebookIcon, 
-    InstagramIcon,
-    TwitterIcon,
-    YouTubeIcon,
-    GitHubIcon,
-    LinkedInIcon,
-    DiscordIcon,
-    TelegramIcon,
     UserInfo,
     JobButton
 } from "./styles/CompanyPortalStyles"
 import LogoImage from "../assets/imgs/logotipo-icone-extendida.svg"
-import LogoImageBlue from "../assets/imgs/logo-icon-name-h.svg"
-import GooglePlayImage from "../assets/imgs/googlePlay.svg"
 import ArrowImage from "../assets/imgs/arrow.svg"
 import PlusSignImage from "../assets/imgs/plus-sign.png"
+import Footer from '../components/CompanyPortal/Footer';
 import { useNavigate } from "react-router-dom"
 import { AuthContext } from "../contexts/Auth/AuthContext"
 import { useContext, useEffect, useState } from "react"
@@ -95,85 +84,7 @@ const CompanyPortal = () => {
         <Main>
             <Row/>
         </Main>
-        <Footer>
-            <div>
-                <img src={LogoImageBlue} width="194px" height="29px" alt="logo" />
-            </div>
-            <div>
-                <h3>Institucional</h3>
-                <ul>
-                    <li>
-                        <a href="https://www.soujunior.tech/">Site SouJunior</a>
-                    </li>
-                    <li>
-                        <a href="https://mentores.soujunior.tech">Portal de mentoria</a>
-                    </li>
-                </ul>
-            </div>
-            <div>
-                <h3>Empresa</h3>
-                <ul>
-                    <li>
-                        <a href="/">Home</a>
-                    </li>
-                </ul>
-            </div>
-            <div>
-            <h3>Aplicativo</h3>
-                <ul>
-                    <li>
-                        <a href="/">
-                            <img src={GooglePlayImage} alt="google play" />
-                        </a>
-                    </li>
-                </ul>
-            </div>
-            <div>
-                <h3>Redes sociais</h3>
-                <Grid>
-                    <li>
-                        <a href="https://www.facebook.com/people/SouJunior/100086671131030/">
-                            <FacebookIcon/>
-                        </a>
-                    </li>
-                    <li>
-                        <a href="https://www.instagram.com/soujunior.tech/">
-                            <InstagramIcon/>
-                        </a>
-                    </li>
-                    <li>
-                        <a href="https://twitter.com/SouJunior_Tech">
-                            <TwitterIcon/>
-                        </a>
-                    </li>
-                    <li>
-                        <a href="https://www.youtube.com/@soujuniortech">
-                            <YouTubeIcon/>
-                        </a>
-                    </li>
-                    <li>
-                        <a href="https://github.com/SouJunior">
-                            <GitHubIcon/>
-                        </a>
-                    </li>
-                    <li>
-                        <a href="https://www.linkedin.com/company/soujunior/">
-                            <LinkedInIcon/>
-                        </a>
-                    </li>
-                    <li>
-                        <a href="https://discord.com/invite/soujunior-community-759176734460346423">
-                            <DiscordIcon/>
-                        </a>
-                    </li>
-                    <li>
-                        <a href="https://t.me/soujuniortech">
-                            <TelegramIcon/>
-                        </a>
-                    </li>
-                </Grid>
-            </div>
-        </Footer>
+        <Footer />
         <Main>
             <Row/>
         </Main>
